requirements.txt
yt-dlp
ffmpeg-python
beautifulsoup4
requests
